Change links to fluentbit.io website
The documentation for the Fluent Bit Loki output plugin references the
fluentbit.io website multiple times. However, the structure of this
website seems to have changed. This commit updates the links to the
correct URLs (if possible).

Additionally, renamed occurences of "fluentbit" to "Fluent Bit", which
seems to be the correct spelling of the product according to their
website.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>